Fix parsing of quoted strings
Pretty printing was not distinguishing between single-quote delimited strings and double-quote delimited strings, which in earlier revisions caused an exception which the latest revision caught and swallowed, but didn't really fix. This change refactors the prettyPrint method as a state machine which correctly handles both quoting and escaping within strings. The isEscaped method is eliminated.