Adapt json-assert and json-path-web-test projects
The json-path allows now to specify whether the parsing of a json argument must be done strictly or not. The strict mode means that a valid JSON data structure is expected (JSONObject or JSONArray). If it is no the case an InvalidJsonException is thrown. Otherwise the parse method can return any valid JSON value (string, numeric)