2.3 merge update: Also fix for quote evaluation
This is a problem in JsonPath where the left-hand side string value was being
quoted by mistake.

Issue 410 on github: https://github.com/json-path/JsonPath/issues/

Also fix issue 409 as well. This should improve performance a bit when parsing.

There are casts because the compiler in IntelliJ was confused as to what overloaded
method to call. Casts don't really hurt.