Automatic module names for Java 9+ modules
'assert' is a reserved keyword in Java and indeed also in
module-info.java (Java module system), therefore the automatic
jar-to-module convert fails and when using the library with a Java
module, the compiler can not find the classes, as the module name
errors.

This added the standardized attribute in the manifest for naming the
module, and does not modularize the code as such. It does allow the
library to be used from Java 9+ modules.

There should be no effect on Java 8 and below.