When the path parameters to a length function for $..book.length() is translated to $.length($..book) its really taking the length of the structure pointed at by the result set $..book not $..book.* (children of book).
This was caused when the CompiledPath#invertScannerFunctionRelationship was introduced, effectively in the graph of CompiledPath translating the $..book.length() to $.length($..book) but what it should have done (for length only) is translated it to $.length($..book.*).  Rather than make this bugfix in the CompiledPath source code I've made it to the Length source as this shouldn't be an issue for any other routine - length is specific in that its asking about the children (hence the .length() the dot there indicates next node(s)).

This addresses ticket 650 - https://github.com/json-path/JsonPath/issues/650 and adds unit tests for the same.