Only look for PatternFlags after a regular expression
As described in #658, the current implementation doesn't work on a logical expression after
an regex like

  ?(@ =~ /something/ || @ == "abc")

This issue has been introduced with the handling of PatternFlags in cbdc9c82e00edb9496f41a4aca3d12ef7ee8321b
The parser should only look for valid PatternFlag chars after the closing slash. It would be nice,
if the PatternFlag enum could return the available flags as chars (or even better a set of chars?). That would
make looking for the next non-flag char easier. (Maybe even add a method to CharacterIndex to find the next
char not in a given set?)