Fix parsing logic of function parameters with closing parentheses
Currently a function with parameter which have parentheses is wrongly parsed. As soon as a closing parentheses is encountered, the parsing logic assumes end state for the function parameter. For example, given 
`concat(sum(), 10)` then `sum(` is treated as first parameter. This is wrong parsing logic.

A parameter should be treated completely parsed only when a `,` or, the last `)` is encountered ( that is groupParen == 0)